refactor(studio): align the auth work with repo conventions
- No barrel files: src/auth/index.ts becomes create.ts, and the type
  re-export goes so consumers import from service.ts, where the types
  are defined.
- Reuse fresco-ui: AppLayout's sign-out error uses Alert instead of a
  raw span with role="alert".
- Don't gate tests on a dependency they don't use: the upload
  authorisation cases return from middleware before the store is
  consulted, so they no longer skip without MinIO. The retrieval case
  does reach store.get and stays gated.
- Accessibility: hide the decorative "or" divider from assistive
  technology.
- Clean lint and knip: drop a redundant knip entry and an
  oxlint-disable .oxlintrc already covers, and narrow values.email
  rather than String()-ing it.
