+++ apps/studio/server/src/auth/create.ts
@@ -6,8 +6,6 @@ import { createBetterAuthService } from './better-auth.ts';
 import { createMailer } from './email.ts';
 import { type AuthService, createDisabledAuthService } from './service.ts';
 
-export type { AuthService, Principal, SessionPrincipal } from './service.ts';
-
 export function createAuthService(env: StudioEnv, pool?: pg.Pool): AuthService {
   if (!env.db || !env.auth) return createDisabledAuthService();
   return createBetterAuthService(
